massive overhaul to the UI + Search and match Capabilities of Angular App. gave this all of the features of the Angular app but included some animations and fun stuff.
I could break this into two commits or three even, to distribute the changes to be one commit per ticket/change. This is how i have learned and prefer to do commits in a professional environment, and i would have done this in that way if this was not just my own little project where i am trying to get the code done with and the integrity of the commits and what we are pushing up is trivial.

ui updates, bootstrap, added all search and features to match that of the angular project capabilities.